+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- DomainBlocker class to manage domain blocks.
+
 ## [0.49.1] - 2026-06-28
 
 ### Added